requirements.txt
alembic==1.16.4
python-dotenv==1.0.0
Flask==2.3.3
SQLAlchemy==2.0.42
python-telegram-bot==20.7
gunicorn==21.2.0
apscheduler==3.10.4
pytest==8.2.1
flake8==7.0.0
black==24.4.2
isort==5.13.2
requests==2.32.3
psycopg[binary]==3.2.9
openai>=1.0.0
